Generates 10 random numbers asynchronously using list comprehension.

+++ 0x02-python_async_comprehension/1-async_comprehension.py
@@ -8,7 +8,8 @@ from typing import Generator
 
 async_generator = __import__('0-async_generator').async_generator
 
-async def async_comprehension() -> list[int]:
+
+async def async_comprehension() -> list[float]:
     """Generates 10 random numbers asynchronously using list comprehension."""
     result = [i async for i in async_generator()]
     return result
